relocate tests for cell and markdown preview
